COMP: Had to cast to double to avoid ambigous call to overloaded func

+++ VTK/Common/vtkFunctionParser.cxx
@@ -540,7 +540,8 @@ void vtkFunctionParser::Evaluate()
           vtkErrorMacro("Trying to take a logarithm of a negative value");
           return;
           }
-        this->Stack[stackPosition] = log(this->Stack[stackPosition])/log(10);
+        this->Stack[stackPosition] = 
+          log(this->Stack[stackPosition])/log((double)10);
         break;
       case VTK_PARSER_SQUARE_ROOT:
         if (this->Stack[stackPosition] < 0)
